+++ spec/integration/braintree/transaction_gateway_spec.js
@@ -39,6 +39,129 @@ describe('TransactionGateway', function () {
       });
     });
 
+    it('charges a card with billing and shipping address specified', function (done) {
+      let transactionParams = {
+        amount: '5.00',
+        creditCard: {
+          number: '5105105105105100',
+          expirationDate: '05/12'
+        },
+        billing: {
+          streetAddress: '123 Fake St',
+          extendedAddress: 'Suite 403',
+          locality: 'Chicago',
+          region: 'IL',
+          postalCode: '60607',
+          phoneNumber: '122-555-1237',
+          countryName: 'United States of America'
+        },
+        shipping: {
+          streetAddress: '456 W Main St',
+          extendedAddress: 'Apt 2F',
+          locality: 'Bartlett',
+          region: 'IL',
+          phoneNumber: '122-555-1236',
+          postalCode: '60103',
+          countryName: 'Mexico',
+          shippingMethod: 'electronic'
+        }
+      };
+
+      specHelper.defaultGateway.transaction.sale(transactionParams, function (err, response) {
+        assert.isNull(err);
+        assert.isTrue(response.success);
+        assert.equal(response.transaction.type, 'sale');
+        assert.equal(response.transaction.amount, '5.00');
+        assert.equal(response.transaction.creditCard.maskedNumber, '510510******5100');
+        assert.isNull(response.transaction.voiceReferralNumber);
+        assert.equal(response.transaction.processorResponseCode, '1000');
+        assert.equal(response.transaction.processorResponseType, 'approved');
+        assert.exists(response.transaction.authorizationExpiresAt);
+
+        done();
+      });
+    });
+
+    it('handles an error when shipping phone number is invalid', function (done) {
+      let transactionParams = {
+        type: 'sale',
+        amount: '64.05',
+        paymentMethodNonce: Nonces.AbstractTransactable,
+        shipping: {
+          streetAddress: '456 W Main St',
+          extendedAddress: 'Apt 2F',
+          locality: 'Bartlett',
+          region: 'IL',
+          phoneNumber: '123-234-3456=098765',
+          postalCode: '60103',
+          countryName: 'Mexico',
+          shippingMethod: 'electronic'
+        }
+      };
+
+      specHelper.defaultGateway.transaction.sale(transactionParams, function (err, response) {
+        assert.isFalse(response.success, 'response had no errors');
+        assert.equal(
+          response.errors.for('transaction').for('shipping').on('phoneNumber')[0].code,
+          ValidationErrorCodes.Transaction.ShippingPhoneNumberIsInvalid
+        );
+        done();
+      });
+    });
+
+    it('handles an error when shipping method is invalid', function (done) {
+      let transactionParams = {
+        type: 'sale',
+        amount: '64.05',
+        paymentMethodNonce: Nonces.AbstractTransactable,
+        shipping: {
+          streetAddress: '456 W Main St',
+          extendedAddress: 'Apt 2F',
+          locality: 'Bartlett',
+          region: 'IL',
+          phoneNumber: '122-555-1236',
+          postalCode: '60103',
+          countryName: 'Mexico',
+          shippingMethod: 'urgent'
+        }
+      };
+
+      specHelper.defaultGateway.transaction.sale(transactionParams, function (err, response) {
+        assert.isFalse(response.success, 'response had no errors');
+        assert.equal(
+          response.errors.for('transaction').for('shipping').on('shippingMethod')[0].code,
+          ValidationErrorCodes.Transaction.ShippingMethodIsInvalid
+        );
+        done();
+      });
+    });
+
+    it('handles an error when billing phone number is invalid', function (done) {
+      let transactionParams = {
+        type: 'sale',
+        amount: '64.05',
+        paymentMethodNonce: Nonces.AbstractTransactable,
+        billing: {
+          streetAddress: '456 W Main St',
+          extendedAddress: 'Apt 2F',
+          locality: 'Bartlett',
+          region: 'IL',
+          phoneNumber: '123-234-3456=098765',
+          postalCode: '60103',
+          countryName: 'Mexico'
+        }
+      };
+
+      specHelper.defaultGateway.transaction.sale(transactionParams, function (err, response) {
+        assert.isFalse(response.success, 'response had no errors');
+        assert.equal(
+          response.errors.for('transaction').for('billing').on('phoneNumber')[0].code,
+          ValidationErrorCodes.Transaction.BillingPhoneNumberIsInvalid
+        );
+        done();
+      });
+    });
+
     it('charges an elo card', function (done) {
       let transactionParams = {
         merchantAccountId: 'adyen_ma',
@@ -3261,13 +3384,50 @@ describe('TransactionGateway', function () {
         },
         riskData: {
           customerBrowser: 'Edge',
-          customerIp: '127.0.0.0'
+          customerDeviceId: 'customer_device_id_012',
+          customerIp: '127.0.0.0',
+          customerLocationZip: '91244',
+          customerTenure: 20
         }
       };
 
       specHelper.defaultGateway.transaction.sale(transactionParams, function (err, response) {
         assert.isNull(err);
         assert.isTrue(response.success);
+        done();
+      });
+    });
+
+    it('handles risk data validation errors', function (done) {
+      let transactionParams = {
+        amount: '10.0',
+        creditCard: {
+          number: '5105105105105100',
+          expirationDate: '05/16'
+        },
+        riskData: {
+          customerBrowser: 'Edge' + '1'.repeat(400),
+          customerDeviceId: 'customer_device_id_012' + '3'.repeat(300),
+          customerIp: '127.0.0.0',
+          customerLocationZip: '912$4',
+          customerTenure: '20'
+        }
+      };
+
+      specHelper.defaultGateway.transaction.sale(transactionParams, function (err, response) {
+        assert.isFalse(response.success, 'response had no errors');
+        assert.equal(
+          response.errors.for('transaction').for('riskData').on('customerBrowser')[0].code,
+          ValidationErrorCodes.RiskData.CustomerBrowserIsTooLong
+        );
+        assert.equal(
+          response.errors.for('transaction').for('riskData').on('customerDeviceId')[0].code,
+          ValidationErrorCodes.RiskData.CustomerDeviceIdIsTooLong
+        );
+        assert.equal(
+          response.errors.for('transaction').for('riskData').on('customerLocationZip')[0].code,
+          ValidationErrorCodes.RiskData.CustomerLocationZipInvalidCharacters
+        );
         done();
       });
     });